requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
discord.py==2.3.2
distlib==0.3.5
dnspython==2.3.0
dotenv==0.0.5
filelock==3.8.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
platformdirs==2.5.2
prettytable==3.8.0
pymongo==4.3.3
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
soupsieve==2.4.1
urllib3==2.0.4
virtualenv==20.16.3
wcwidth==0.2.6
yarl==1.9.2
